Moves Image component into ui subdirectory
Reorganizes the codebase by relocating the Image component into a ui subdirectory to better separate UI primitives from higher-level components. Updates all import sites to the new path to avoid broken imports and removes an unused import in the stripe image module.

Improves project structure and clarity without altering runtime behavior.
