+++ src/lib/components/layout/Header.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import ThemeToggle from '$lib/components/theme/ThemeToggle.svelte';
-	import Image from '$lib/components/Image.svelte';
+	import Image from '$lib/components/ui/Image.svelte';
 	import { logo_small_clear, name, pronouns } from '$lib/store';
 	import Slogan from '$lib/components/blocks/Slogan.svelte';
