+++ src/routes/+error.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { page } from '$app/stores';
-	import Image from '$lib/components/Image.svelte';
+	import Image from '$lib/components/ui/Image.svelte';
 </script>
 
 <div class="flex min-h-screen items-center justify-center container py-5">
